+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New command "Merge arbitrary files"
 
 ## [0.1.0] - 2020-08-27
 ### Added